services/requirements-twitter.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
virtuals_tweepy>=0.1.6

